Add RayCast to Player
Create Private Variable Reach and multiply it to the player rotation
vector to extend the reach in the 'look' direction of the player.

The ray cast endpoint  is calculated by adding the vector from origin to
the player location  and the magnitude of the linetrace multiplied by
the playerRotation vector. The raycast is started from player location
and it then ends at the calculated end point.