Provide Input Key Bindings
Input Key Binding Options located in project settings: engine. Added a
Grab action mapping for left shift and right mouse.

Grabber.h: added a private pointer for the  UInputComponent

Grabber.cpp:  Check if InputComponent exists and log it to the console.
from beginplay. Then Bind the inputcomponent to the action mapping.

Created a UGrabber::Grab() method: logs that the action key was pressed.